+++ bimmer_connected/vehicle/doors_windows.py
@@ -4,7 +4,7 @@ import logging
 from dataclasses import dataclass, field
 from typing import Any, Dict, List
 
-from bimmer_connected.vehicle.models import StrEnum, VehicleDataBase
+from bimmer_connected.models import StrEnum, VehicleDataBase
 
 _LOGGER = logging.getLogger(__name__)
