+++ test/test_account.py
@@ -21,7 +21,7 @@ from bimmer_connected.account import ConnectedDriveAccount, MyBMWAccount
 from bimmer_connected.api.authentication import MyBMWAuthentication, MyBMWLoginRetry
 from bimmer_connected.api.regions import get_region_from_name
 from bimmer_connected.const import Regions
-from bimmer_connected.vehicle.models import GPSPosition
+from bimmer_connected.models import GPSPosition
 
 from . import (
     RESPONSE_DIR,
